[refactor] Inputcontainer, ListContainer 로 관심사 분리
